Should use fully qualified item names when looking
Closed #389
